simplified recurrent state tests
